% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn.R
\name{burn}
\alias{burn}
\title{Core Burn function for Andison Fire Module}
\usage{
burn(landscape, startCells, fireSizes = 5, nActiveCells1 = c(10, 36),
  spawnNewActive = c(0.46, 0.2, 0.26, 0.11), sizeCutoffs = c(8000, 20000))
}
\arguments{
\item{landscape}{A RasterLayer. This only provides the extent and resolution for the fire spread algorithm.}

\item{startCells}{A numeric vector indicating the indices on the \code{landscape} where the fires 
will start with 100\% certainty.}

\item{fireSizes}{A numeric vector indicating the final size of each of the fires. Must be same length as
\code{startCells}.}

\item{nActiveCells1}{A numeric vector of length 2. These are cutoffs above and below each of which different values
of \code{spawnNewActive} are used. See details.}

\item{spawnNewActive}{A numeric vector of length 4. These are the probabilities of creating spreading to 2 neighbours
instead of the 1 default neighbour, each time step. The 4 values are for 4 different fire
size conditions. See details.}

\item{sizeCutoffs}{A numeric vector of length 2. These are 2 size thresholds that affect which 
\code{spawnNewActive} probabilities are used. See details.}
}
\value{
A \code{data.table} with 4 columns
}
\description{
The main function for the Andison Fire Module. See details.
}
\details{
This algorithm is a modified contagious cellular automaton. The algorithm is as follows:
}
\section{Algorithm}{


\subsection{Core}{
Each fire starts at a single pixel, \code{startCells} and will spread, i.e,. visit and convert 
from a 0 to the fire id number. It will iteratively 
spread until the number of cells visited is equal to \code{floor(fireSizes)}.
}
\subsection{Adjustments due to current fire size and number of active pixels}{

That can vary too, but it gets a bit complicated, so use that for now.
Spawning probability was originally set at 13%, but created problems with very large and very small fires, so over time has been adjusted to vary depending on a) number of active “firelets” (NF) and b) fire size (FS):
  -         If NF>=10 and <36 and FS <20,000 ha then P = 20%
  -         If NF>36 and FS <8,000 ha, P=11%
  -         If  NF<36 and FS >20,000 ha, P=26%
  -         If NF<10 then P=46%
These rule create more heterogeneity in the pattern of burning.
}
}

\section{Not yet implemented}{

Chances of fire spread stopping at any given pixel in any given firelet is 23%.    
The fire can also get stuck when it turns back in on itself and burns islands, 
but now has no way to grow larger from the perimeter.  
If the fire has not reached its target size, I create new firelet “spot” fires up to 400 pixels away.
}

